Refactor endpoint configuration and error handling structure
