+++ src/OrderPoint.Api/Configuration/IEndpoint.cs
@@ -1,0 +1,6 @@
+﻿namespace OrderPoint.Api.Configuration;
+
+internal interface IEndpoint
+{
+    void MapEndpoint(IEndpointRouteBuilder app);
+}
