+++ src/OrderPoint.Api/Endpoints/Categories/GetCategoriesEndpoint.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using Microsoft.AspNetCore.Mvc;
+using OrderPoint.Api.Configuration;
 using OrderPoint.Api.Extensions;
 using OrderPoint.Application.Dtos;
 using OrderPoint.Application.Queries;
