+++ src/OrderPoint.Api/Program.cs
@@ -1,5 +1,6 @@
 using System.Reflection;
 using FluentValidation;
+using OrderPoint.Api.Configuration;
 using OrderPoint.Api.Exceptions;
 using OrderPoint.Api.Extensions;
 using OrderPoint.Application;
