+++ src/OrderPoint.Domain/Outcomes/ErrorType.cs
@@ -1,0 +1,10 @@
+﻿namespace OrderPoint.Domain.Outcomes;
+
+public enum ErrorType
+{
+    Failure,
+    RequestValidation,
+    Validation,
+    NotFound,
+    Conflict
+}
